ours/requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
networkx>=3.1
openai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
rank-bm25>=0.2.2
nltk>=3.8.1
rouge-score>=0.1.2
bert-score>=0.3.13
tqdm>=4.66.0
cross-encoder-ms-marco-MiniLM-L-6-v2>=0.0.0; python_version >= "0"
